docs: adjust html stackblitz
