+++ packages/components/public/assets/code/html/index.html
@@ -1,7 +1,7 @@
 <html>
   <body>
     <bal-app>
-      <main class="container py-large">
+      <main class="container">
         <bal-heading>Hello World</bal-heading>
       </main>
     </bal-app>
