+++ packages/components/public/assets/code/html/index.ts
@@ -1,1 +1,215 @@
-import './loader';
+/**
+ * Stackblitz Initialization
+ * ------------------------------------
+ * The index.ts file is a custom Stackblitz initialization file.
+ * Since Stackblitz does not support dynamic loading of Stencil components.
+ */
+
+import '@baloise/design-system-components/dist/design-system-components/design-system-components.css';
+import '@baloise/design-system-components/dist/components';
+import { defineCustomElement as defineBalAccordion } from '@baloise/design-system-components/dist/components/bal-accordion';
+import { defineCustomElement as defineBalApp } from '@baloise/design-system-components/dist/components/bal-app';
+import { defineCustomElement as defineBalBadge } from '@baloise/design-system-components/dist/components/bal-badge';
+import { defineCustomElement as defineBalButton } from '@baloise/design-system-components/dist/components/bal-button';
+import { defineCustomElement as defineBalButtonGroup } from '@baloise/design-system-components/dist/components/bal-button-group';
+import { defineCustomElement as defineBalCard } from '@baloise/design-system-components/dist/components/bal-card';
+import { defineCustomElement as defineBalCardActions } from '@baloise/design-system-components/dist/components/bal-card-actions';
+import { defineCustomElement as defineBalCardButton } from '@baloise/design-system-components/dist/components/bal-card-button';
+import { defineCustomElement as defineBalCardContent } from '@baloise/design-system-components/dist/components/bal-card-content';
+import { defineCustomElement as defineBalCardSubtitle } from '@baloise/design-system-components/dist/components/bal-card-subtitle';
+import { defineCustomElement as defineBalCardTitle } from '@baloise/design-system-components/dist/components/bal-card-title';
+import { defineCustomElement as defineBalCarousel } from '@baloise/design-system-components/dist/components/bal-carousel';
+import { defineCustomElement as defineBalCarouselItem } from '@baloise/design-system-components/dist/components/bal-carousel-item';
+import { defineCustomElement as defineBalCheckbox } from '@baloise/design-system-components/dist/components/bal-checkbox';
+import { defineCustomElement as defineBalCheckboxGroup } from '@baloise/design-system-components/dist/components/bal-checkbox-group';
+import { defineCustomElement as defineBalClose } from '@baloise/design-system-components/dist/components/bal-close';
+import { defineCustomElement as defineBalData } from '@baloise/design-system-components/dist/components/bal-data';
+import { defineCustomElement as defineBalDataItem } from '@baloise/design-system-components/dist/components/bal-data-item';
+import { defineCustomElement as defineBalDataLabel } from '@baloise/design-system-components/dist/components/bal-data-label';
+import { defineCustomElement as defineBalDataValue } from '@baloise/design-system-components/dist/components/bal-data-value';
+import { defineCustomElement as defineBalDatepicker } from '@baloise/design-system-components/dist/components/bal-datepicker';
+import { defineCustomElement as defineBalField } from '@baloise/design-system-components/dist/components/bal-field';
+import { defineCustomElement as defineBalFieldControl } from '@baloise/design-system-components/dist/components/bal-field-control';
+import { defineCustomElement as defineBalFieldHint } from '@baloise/design-system-components/dist/components/bal-field-hint';
+import { defineCustomElement as defineBalFieldLabel } from '@baloise/design-system-components/dist/components/bal-field-label';
+import { defineCustomElement as defineBalFieldMessage } from '@baloise/design-system-components/dist/components/bal-field-message';
+import { defineCustomElement as defineBalFileUpload } from '@baloise/design-system-components/dist/components/bal-file-upload';
+import { defineCustomElement as defineBalFooter } from '@baloise/design-system-components/dist/components/bal-footer';
+import { defineCustomElement as defineBalForm } from '@baloise/design-system-components/dist/components/bal-form';
+import { defineCustomElement as defineBalFormCol } from '@baloise/design-system-components/dist/components/bal-form-col';
+import { defineCustomElement as defineBalFormGrid } from '@baloise/design-system-components/dist/components/bal-form-grid';
+import { defineCustomElement as defineBalHeading } from '@baloise/design-system-components/dist/components/bal-heading';
+import { defineCustomElement as defineBalHint } from '@baloise/design-system-components/dist/components/bal-hint';
+import { defineCustomElement as defineBalHintText } from '@baloise/design-system-components/dist/components/bal-hint-text';
+import { defineCustomElement as defineBalHintTitle } from '@baloise/design-system-components/dist/components/bal-hint-title';
+import { defineCustomElement as defineBalIcon } from '@baloise/design-system-components/dist/components/bal-icon';
+import { defineCustomElement as defineBalInput } from '@baloise/design-system-components/dist/components/bal-input';
+import { defineCustomElement as defineBalInputGroup } from '@baloise/design-system-components/dist/components/bal-input-group';
+import { defineCustomElement as defineBalInputSlider } from '@baloise/design-system-components/dist/components/bal-input-slider';
+import { defineCustomElement as defineBalInputStepper } from '@baloise/design-system-components/dist/components/bal-input-stepper';
+import { defineCustomElement as defineBalList } from '@baloise/design-system-components/dist/components/bal-list';
+import { defineCustomElement as defineBalListItem } from '@baloise/design-system-components/dist/components/bal-list-item';
+import { defineCustomElement as defineBalListItemAccordionBody } from '@baloise/design-system-components/dist/components/bal-list-item-accordion-body';
+import { defineCustomElement as defineBalListItemAccordionHead } from '@baloise/design-system-components/dist/components/bal-list-item-accordion-head';
+import { defineCustomElement as defineBalListItemContent } from '@baloise/design-system-components/dist/components/bal-list-item-content';
+import { defineCustomElement as defineBalListItemIcon } from '@baloise/design-system-components/dist/components/bal-list-item-icon';
+import { defineCustomElement as defineBalListItemSubtitle } from '@baloise/design-system-components/dist/components/bal-list-item-subtitle';
+import { defineCustomElement as defineBalListItemTitle } from '@baloise/design-system-components/dist/components/bal-list-item-title';
+import { defineCustomElement as defineBalLogo } from '@baloise/design-system-components/dist/components/bal-logo';
+import { defineCustomElement as defineBalModal } from '@baloise/design-system-components/dist/components/bal-modal';
+import { defineCustomElement as defineBalModalBody } from '@baloise/design-system-components/dist/components/bal-modal-body';
+import { defineCustomElement as defineBalModalHeader } from '@baloise/design-system-components/dist/components/bal-modal-header';
+import { defineCustomElement as defineBalNavbar } from '@baloise/design-system-components/dist/components/bal-navbar';
+import { defineCustomElement as defineBalNavbarBrand } from '@baloise/design-system-components/dist/components/bal-navbar-brand';
+import { defineCustomElement as defineBalNavbarMenu } from '@baloise/design-system-components/dist/components/bal-navbar-menu';
+import { defineCustomElement as defineBalNavbarMenuEnd } from '@baloise/design-system-components/dist/components/bal-navbar-menu-end';
+import { defineCustomElement as defineBalNavbarMenuStart } from '@baloise/design-system-components/dist/components/bal-navbar-menu-start';
+import { defineCustomElement as defineBalNavigation } from '@baloise/design-system-components/dist/components/bal-navigation';
+import { defineCustomElement as defineBalNavigationLevelBlock } from '@baloise/design-system-components/dist/components/bal-navigation-level-block';
+import { defineCustomElement as defineBalNavigationLevelBlockItem } from '@baloise/design-system-components/dist/components/bal-navigation-level-block-item';
+import { defineCustomElement as defineBalNavigationLevelMain } from '@baloise/design-system-components/dist/components/bal-navigation-level-main';
+import { defineCustomElement as defineBalNavigationLevelMeta } from '@baloise/design-system-components/dist/components/bal-navigation-level-meta';
+import { defineCustomElement as defineBalNavigationLevels } from '@baloise/design-system-components/dist/components/bal-navigation-levels';
+import { defineCustomElement as defineBalNavigationMain } from '@baloise/design-system-components/dist/components/bal-navigation-main';
+import { defineCustomElement as defineBalNavigationMainBody } from '@baloise/design-system-components/dist/components/bal-navigation-main-body';
+import { defineCustomElement as defineBalNavigationMainHead } from '@baloise/design-system-components/dist/components/bal-navigation-main-head';
+import { defineCustomElement as defineBalNavigationMenu } from '@baloise/design-system-components/dist/components/bal-navigation-menu';
+import { defineCustomElement as defineBalNavigationMenuList } from '@baloise/design-system-components/dist/components/bal-navigation-menu-list';
+import { defineCustomElement as defineBalNavigationMenuListItem } from '@baloise/design-system-components/dist/components/bal-navigation-menu-list-item';
+import { defineCustomElement as defineBalNavigationMeta } from '@baloise/design-system-components/dist/components/bal-navigation-meta';
+import { defineCustomElement as defineBalNavigationMetaEnd } from '@baloise/design-system-components/dist/components/bal-navigation-meta-end';
+import { defineCustomElement as defineBalNavigationMetaStart } from '@baloise/design-system-components/dist/components/bal-navigation-meta-start';
+import { defineCustomElement as defineBalNavigationPopover } from '@baloise/design-system-components/dist/components/bal-navigation-popover';
+import { defineCustomElement as defineBalNotices } from '@baloise/design-system-components/dist/components/bal-notices';
+import { defineCustomElement as defineBalNotification } from '@baloise/design-system-components/dist/components/bal-notification';
+import { defineCustomElement as defineBalNumberInput } from '@baloise/design-system-components/dist/components/bal-number-input';
+import { defineCustomElement as defineBalPagination } from '@baloise/design-system-components/dist/components/bal-pagination';
+import { defineCustomElement as defineBalPopover } from '@baloise/design-system-components/dist/components/bal-popover';
+import { defineCustomElement as defineBalPopoverContent } from '@baloise/design-system-components/dist/components/bal-popover-content';
+import { defineCustomElement as defineBalRadio } from '@baloise/design-system-components/dist/components/bal-radio';
+import { defineCustomElement as defineBalRadioGroup } from '@baloise/design-system-components/dist/components/bal-radio-group';
+import { defineCustomElement as defineBalSelect } from '@baloise/design-system-components/dist/components/bal-select';
+import { defineCustomElement as defineBalSelectOption } from '@baloise/design-system-components/dist/components/bal-select-option';
+import { defineCustomElement as defineBalShape } from '@baloise/design-system-components/dist/components/bal-shape';
+import { defineCustomElement as defineBalSheet } from '@baloise/design-system-components/dist/components/bal-sheet';
+import { defineCustomElement as defineBalSnackbar } from '@baloise/design-system-components/dist/components/bal-snackbar';
+import { defineCustomElement as defineBalSpinner } from '@baloise/design-system-components/dist/components/bal-spinner';
+import { defineCustomElement as defineBalStage } from '@baloise/design-system-components/dist/components/bal-stage';
+import { defineCustomElement as defineBalStageBackLink } from '@baloise/design-system-components/dist/components/bal-stage-back-link';
+import { defineCustomElement as defineBalStageBody } from '@baloise/design-system-components/dist/components/bal-stage-body';
+import { defineCustomElement as defineBalStageFoot } from '@baloise/design-system-components/dist/components/bal-stage-foot';
+import { defineCustomElement as defineBalStageHead } from '@baloise/design-system-components/dist/components/bal-stage-head';
+import { defineCustomElement as defineBalStageImage } from '@baloise/design-system-components/dist/components/bal-stage-image';
+import { defineCustomElement as defineBalTabItem } from '@baloise/design-system-components/dist/components/bal-tab-item';
+import { defineCustomElement as defineBalTable } from '@baloise/design-system-components/dist/components/bal-table';
+import { defineCustomElement as defineBalTabs } from '@baloise/design-system-components/dist/components/bal-tabs';
+import { defineCustomElement as defineBalTag } from '@baloise/design-system-components/dist/components/bal-tag';
+import { defineCustomElement as defineBalTagGroup } from '@baloise/design-system-components/dist/components/bal-tag-group';
+import { defineCustomElement as defineBalText } from '@baloise/design-system-components/dist/components/bal-text';
+import { defineCustomElement as defineBalTextarea } from '@baloise/design-system-components/dist/components/bal-textarea';
+import { defineCustomElement as defineBalToast } from '@baloise/design-system-components/dist/components/bal-toast';
+
+import './example';
+
+defineBalAccordion();
+defineBalApp();
+defineBalBadge();
+defineBalButton();
+defineBalButtonGroup();
+defineBalCard();
+defineBalCardActions();
+defineBalCardButton();
+defineBalCardContent();
+defineBalCardSubtitle();
+defineBalCardTitle();
+defineBalCarousel();
+defineBalCarouselItem();
+defineBalCheckbox();
+defineBalCheckboxGroup();
+defineBalClose();
+defineBalData();
+defineBalDataItem();
+defineBalDataLabel();
+defineBalDataValue();
+defineBalDatepicker();
+defineBalField();
+defineBalFieldControl();
+defineBalFieldHint();
+defineBalFieldLabel();
+defineBalFieldMessage();
+defineBalFileUpload();
+defineBalFooter();
+defineBalForm();
+defineBalFormCol();
+defineBalFormGrid();
+defineBalHeading();
+defineBalHint();
+defineBalHintText();
+defineBalHintTitle();
+defineBalIcon();
+defineBalInput();
+defineBalInputGroup();
+defineBalInputSlider();
+defineBalInputStepper();
+defineBalList();
+defineBalListItem();
+defineBalListItemAccordionBody();
+defineBalListItemAccordionHead();
+defineBalListItemContent();
+defineBalListItemIcon();
+defineBalListItemSubtitle();
+defineBalListItemTitle();
+defineBalLogo();
+defineBalModal();
+defineBalModalBody();
+defineBalModalHeader();
+defineBalNavbar();
+defineBalNavbarBrand();
+defineBalNavbarMenu();
+defineBalNavbarMenuEnd();
+defineBalNavbarMenuStart();
+defineBalNavigation();
+defineBalNavigationLevelBlock();
+defineBalNavigationLevelBlockItem();
+defineBalNavigationLevelMain();
+defineBalNavigationLevelMeta();
+defineBalNavigationLevels();
+defineBalNavigationMain();
+defineBalNavigationMainBody();
+defineBalNavigationMainHead();
+defineBalNavigationMenu();
+defineBalNavigationMenuList();
+defineBalNavigationMenuListItem();
+defineBalNavigationMeta();
+defineBalNavigationMetaEnd();
+defineBalNavigationMetaStart();
+defineBalNavigationPopover();
+defineBalNotices();
+defineBalNotification();
+defineBalNumberInput();
+defineBalPagination();
+defineBalPopover();
+defineBalPopoverContent();
+defineBalRadio();
+defineBalRadioGroup();
+defineBalSelect();
+defineBalSelectOption();
+defineBalShape();
+defineBalSheet();
+defineBalSnackbar();
+defineBalSpinner();
+defineBalStage();
+defineBalStageBackLink();
+defineBalStageBody();
+defineBalStageFoot();
+defineBalStageHead();
+defineBalStageImage();
+defineBalTabItem();
+defineBalTable();
+defineBalTabs();
+defineBalTag();
+defineBalTagGroup();
+defineBalText();
+defineBalTextarea();
+defineBalToast();
+
