+++ packages/components/src/components/bal-accordion/stories/usage/html/example_ts.md
@@ -1,3 +1,2 @@
 ```ts
-import './loader'
 ```
